@@ -1,44 +1,43 @@
-import { Component } from 'react';
+import { useEffect } from 'react';
 import { createPortal} from 'react-dom'
 import PropTypes from 'prop-types';
 import './Modal.module.css';
 
 const modalRoot = document.querySelector('#modal-root');
 
- export class Modal extends Component {
-  componentDidMount() {
-    window.addEventListener('keydown', this.handleKeyDown);
+export const Modal = ({ largeImageURL, onClose }) => {
+  useEffect(() => {
+    window.addEventListener('keydown', handleKeyDown);
+   return () => {
+      window.removeEventListener('keydown', handleKeyDown);
   }
-
-  componentWillUnmount() {
-    window.removeEventListener('keydown', this.handleKeyDown);
-  }
-
-  handleKeyDown = e => {
+  });
+  
+ 
+  const handleKeyDown = e => {
     if (e.code === 'Escape') {
-      this.props.onClose();
+      onClose();
     }
   };
 
-  handleBackDropClick = e => {
+  const handleBackDropClick = e => {
     if (e.currentTarget === e.target) {
-      this.props.onClose();
+      onClose();
     }
   };
 
-  render() {
     return createPortal(
-      <div className='Overlay' onClick={this.handleBackDropClick}>
+      <div className='Overlay' onClick={handleBackDropClick}>
         <div className='Modal'>
-          <img src={this.props.largeImageURL} alt="" />
+          <img src={largeImageURL} alt="" />
         </div>
       </div>,
       modalRoot
     );
   }
-}
+
 
 Modal.propTypes = {
-  onClose: PropTypes.func.isRequired,
-  largeImageUrl: PropTypes.string.isRequired,
+  onClose: PropTypes.func,
+  largeImageUrl: PropTypes.string,
 };
