@@ -1,29 +1,26 @@
-import React, { Component } from 'react';
+import { useState } from 'react';
 import './Searchbar.module.css';
 import PropTypes from 'prop-types';
 
-export class Searchbar extends Component {
-  state = {
-    searchQuery: '',
+export const Searchbar = ({ onSubmit}) => {
+ const [searchQuery, setSearchQuery] = useState('');
+
+ const handleChange = e => {
+    setSearchQuery(e.target.value.toLowerCase());
   };
 
-  handleChange = e => {
-    this.setState({ searchQuery: e.currentTarget.value.toLowerCase() });
-  };
-
-  handleSubmit = e => {
+ const handleSubmit = e => {
     e.preventDefault();
-    if (this.state.searchQuery.trim() === '') {
-      return alert('Please enter something :)');
+    if (searchQuery.trim() === '') {
+      return alert('Please enter something');
     }
-    this.props.onSubmit(this.state.searchQuery);
-    this.setState({ searchQuery: '' });
+    onSubmit(searchQuery);
+    setSearchQuery('');
   };
 
-  render() {
     return (
       <header>
-        <form onSubmit={this.handleSubmit}>
+        <form onSubmit={handleSubmit}>
           <button type="submit">Search
           </button>
 
@@ -33,14 +30,14 @@ export class Searchbar extends Component {
             autoFocus
             placeholder="Search images and photos"
             name="searchQuery"
-            value={this.state.searchQuery}
-            onChange={this.handleChange}
+            value={searchQuery}
+            onChange={handleChange}
           />
         </form>
       </header>
     );
   }
-}
+
 Searchbar.propTypes = {
   onSubmit: PropTypes.func,
 };
